use env var to control cleanup

+++ .Build.ps1
@@ -237,9 +237,9 @@ Add-BuildTask IntegrationTestAzureVMs {
 .Synopsis: remove all assets deployed to Azure and any local temporary changes (should be none)
 #>
 Exit-Build {
-    # temporary for diagnostics and QA
-    Start-Sleep -Seconds 600
-    Remove-AzureTestResources
+    if (!$env:RemoveRG) {
+        Remove-AzureTestResources
+    }
 }
 
 <#
